refactor: extract core endpoint mapping
